pyproject.toml
[project]
name = "library-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=6.2.1",
    "cryptography>=46.0.2",
    "fastapi[standard]>=0.118.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
]

[build-system]
requires = ["uv_build>=0.8.24,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pyright>=1.1.406",
    "pytest>=8.4.2",
    "pytest-httpx>=0.35.0",
    "ruff>=0.13.3",
]

[project.scripts]
library-api = "library_api.api.kernel:server"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "B",   # flake8-bugbear: https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "D",   # pydocstyle: https://docs.astral.sh/ruff/rules/#pydocstyle-d
    "E",   # pycodestyle errors: https://docs.astral.sh/ruff/rules/#error-e
    "F",   # pyflakes: https://docs.astral.sh/ruff/rules/#pyflakes-f
    "Q",   # flake8-quotes: https://docs.astral.sh/ruff/rules/#flake8-quotes-q
    "ANN", # flake8-annotations: https://docs.astral.sh/ruff/rules/#flake8-annotations-ann
]
ignore = [
    "D203", # incorrect-blank-line-before-class: https://docs.astral.sh/ruff/rules/incorrect-blank-line-before-class/
    "D213"  # multi-line-summary-second-line: https://docs.astral.sh/ruff/rules/multi-line-summary-second-line/
]
